backend/face_service/requirements.txt
flask
flask-cors
deepface
numpy
Pillow 